+++ src/main/kotlin/net/rentalhost/plugins/php/hammer/inspections/deprecation/DollarSignOutsideCurlyBracesInspection.kt
@@ -6,6 +6,8 @@ import com.intellij.psi.PsiElementVisitor
 import com.jetbrains.php.lang.inspections.PhpInspection
 import com.jetbrains.php.lang.psi.elements.impl.StringLiteralExpressionImpl
 import com.jetbrains.php.lang.psi.elements.impl.VariableImpl
+import net.rentalhost.plugins.services.FactoryService
+import net.rentalhost.plugins.services.LocalQuickFixService
 import net.rentalhost.plugins.services.ProblemsHolderService
 
 class DollarSignOutsideCurlyBracesInspection: PhpInspection() {
@@ -19,10 +21,16 @@ class DollarSignOutsideCurlyBracesInspection: PhpInspection() {
 
                     if (elementCurly.startsWith("\${") &&
                         elementCurly.endsWith("}")) {
+                        val replaceWith = elementCurly.substring(2, elementCurly.length - 1)
+
                         ProblemsHolderService.registerProblem(
                             problemsHolder,
                             element,
-                            "Deprecated: using \${var} in strings."
+                            "Deprecated: using \${var} in strings.",
+                            LocalQuickFixService.SimpleReplaceQuickFix(
+                                "Replace with {\$$replaceWith}",
+                                FactoryService.createCurlyVariable(problemsHolder.project, replaceWith)
+                            )
                         )
                     }
                 }