+++ src/main/kotlin/net/rentalhost/plugins/services/FactoryService.kt
@@ -100,4 +100,7 @@ object FactoryService {
 
     fun createTypeCast(project: Project, castType: String): LeafPsiElement =
         PhpPsiElementFactory.createPhpPsiFromText(project, UnaryExpression::class.java, "$castType\$x").firstChild as LeafPsiElement
+
+    fun createCurlyVariable(project: Project, variableName: String): PsiElement =
+        PhpPsiElementFactory.createPhpPsiFromText(project, Variable::class.java, "\"{\$$variableName}\";")
 }